+ Spring MVC

+++ src/main/java/org/simple/twitter/dao/user/UserDao.java
@@ -3,5 +3,5 @@ package org.simple.twitter.dao.user;
 import org.simple.twitter.dao.Dao;
 import org.simple.twitter.model.User;
 
-public interface UserDao extends Dao<User, Long> {
+public interface UserDao extends Dao<User, Integer> {
 }
